requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.7.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylama==8.3.7
python-dotenv==0.19.2
requests==2.27.1
snowballstemmer==2.2.0
urllib3==1.26.8
Werkzeug==2.0.2
